Applied search highlight and scroll animation refactoring for #49457
